Migration mobile Transaction component to TransactionListItem + cleanup
