+++ packages/desktop-client/src/components/mobile/transactions/TransactionList.jsx
@@ -5,8 +5,8 @@ import React, {
   useRef,
   useState,
 } from 'react';
+import { ListBox, Section, Header, Collection } from 'react-aria-components';
 import { useDispatch } from 'react-redux';
-import { Item, Section } from 'react-stately';
 
 import { t } from 'i18next';
 
@@ -34,17 +34,16 @@ import { Menu } from '../../common/Menu';
 import { Popover } from '../../common/Popover';
 import { Text } from '../../common/Text';
 import { View } from '../../common/View';
+import { useScroll } from '../../ScrollProvider';
 import { FloatingActionBar } from '../FloatingActionBar';
 
-import { ListBox } from './ListBox';
-import { Transaction } from './Transaction';
+import { TransactionListItem } from './TransactionListItem';
 
 const NOTIFICATION_BOTTOM_INSET = 75;
 
 export function TransactionList({
   isLoading,
   transactions,
-  isNewTransaction,
   onOpenTransaction,
   onLoadMore,
 }) {
@@ -56,22 +55,14 @@ export function TransactionList({
         sections.length === 0 ||
         transaction.date !== sections[sections.length - 1].date
       ) {
-        // Mark the last transaction in the section so it can render
-        // with a different border
-        const lastSection = sections[sections.length - 1];
-        if (lastSection && lastSection.data.length > 0) {
-          const lastData = lastSection.data;
-          lastData[lastData.length - 1].isLast = true;
-        }
-
         sections.push({
           id: `${isPreviewId(transaction.id) ? 'preview/' : ''}${transaction.date}`,
           date: transaction.date,
-          data: [],
+          transactions: [],
         });
       }
 
-      sections[sections.length - 1].data.push(transaction);
+      sections[sections.length - 1].transactions.push(transaction);
     });
     return sections;
   }, [transactions]);
@@ -79,15 +70,27 @@ export function TransactionList({
   const dispatchSelected = useSelectedDispatch();
   const selectedTransactions = useSelectedItems();
 
-  const onTransactionPress = (transaction, isLongPress = false) => {
-    const isPreview = isPreviewId(transaction.id);
+  const onTransactionPress = useCallback(
+    (transaction, isLongPress = false) => {
+      const isPreview = isPreviewId(transaction.id);
+      if (!isPreview && (isLongPress || selectedTransactions.size > 0)) {
+        dispatchSelected({ type: 'select', id: transaction.id });
+      } else {
+        onOpenTransaction(transaction);
+      }
+    },
+    [dispatchSelected, onOpenTransaction, selectedTransactions.size],
+  );
 
-    if (!isPreview && (isLongPress || selectedTransactions.size > 0)) {
-      dispatchSelected({ type: 'select', id: transaction.id });
-    } else {
-      onOpenTransaction(transaction);
+  const { hasScrolledToBottom } = useScroll();
+  const scrolledToBottom = hasScrolledToBottom(5);
+  const prevScrolledToBottom = useRef(scrolledToBottom);
+
+  useEffect(() => {
+    if (!prevScrolledToBottom.current && scrolledToBottom) {
+      onLoadMore?.();
     }
-  };
+  }, [onLoadMore, scrolledToBottom]);
 
   if (isLoading) {
     return (
@@ -103,65 +106,63 @@ export function TransactionList({
       </View>
     );
   }
-
   return (
     <>
       <ListBox
         aria-label="Transaction list"
-        loadMore={onLoadMore}
-        selectionMode="none"
+        selectionMode={selectedTransactions.size > 0 ? 'multiple' : 'single'}
+        selectedKeys={selectedTransactions}
+        renderEmptyState={() => (
+          <View
+            style={{
+              alignItems: 'center',
+              justifyContent: 'center',
+              backgroundColor: theme.mobilePageBackground,
+            }}
+          >
+            <Text style={{ fontSize: 15 }}>No transactions</Text>
+          </View>
+        )}
+        items={sections}
+        dependencies={[selectedTransactions]}
       >
-        {sections.length === 0 ? (
+        {section => (
           <Section>
-            <Item textValue="No transactions">
-              <div
-                style={{
-                  display: 'flex',
-                  justifyContent: 'center',
-                  width: '100%',
-                  backgroundColor: theme.mobilePageBackground,
-                }}
-              >
-                <Text style={{ fontSize: 15 }}>No transactions</Text>
-              </div>
-            </Item>
-          </Section>
-        ) : null}
-        {sections.map(section => {
-          return (
-            <Section
-              title={
-                <span>{monthUtils.format(section.date, 'MMMM dd, yyyy')}</span>
-              }
-              key={section.id}
+            <Header
+              style={{
+                ...styles.smallText,
+                backgroundColor: theme.pageBackground,
+                borderBottom: `1px solid ${theme.tableBorder}`,
+                borderTop: `1px solid ${theme.tableBorder}`,
+                color: theme.tableHeaderText,
+                display: 'flex',
+                justifyContent: 'center',
+                paddingBottom: 4,
+                paddingTop: 4,
+                position: 'sticky',
+                top: '0',
+                width: '100%',
+                zIndex: 10,
+              }}
             >
-              {section.data.map((transaction, index, transactions) => {
-                if (isPreviewId(transaction.id) && transaction.is_child) {
-                  return null;
-                }
-
-                return (
-                  <Item
-                    key={transaction.id}
-                    style={{
-                      fontSize:
-                        index === transactions.length - 1 ? 98 : 'inherit',
-                    }}
-                    textValue={transaction.id}
-                  >
-                    <Transaction
-                      transaction={transaction}
-                      isAdded={isNewTransaction(transaction.id)}
-                      isSelected={selectedTransactions.has(transaction.id)}
-                      onPress={trans => onTransactionPress(trans)}
-                      onLongPress={trans => onTransactionPress(trans, true)}
-                    />
-                  </Item>
-                );
-              })}
-            </Section>
-          );
-        })}
+              {monthUtils.format(section.date, 'MMMM dd, yyyy')}
+            </Header>
+            <Collection
+              items={section.transactions.filter(
+                t => !isPreviewId(t.id) || !t.is_child,
+              )}
+              addIdAndValue
+            >
+              {transaction => (
+                <TransactionListItem
+                  value={transaction}
+                  onPress={trans => onTransactionPress(trans)}
+                  onLongPress={trans => onTransactionPress(trans, true)}
+                />
+              )}
+            </Collection>
+          </Section>
+        )}
       </ListBox>
       {selectedTransactions.size > 0 && (
         <SelectedTransactionsFloatingActionBar transactions={transactions} />
