+++ packages/desktop-client/src/components/mobile/transactions/TransactionListWithBalances.jsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import { useSelector } from 'react-redux';
 
 import { t } from 'i18next';
 
@@ -70,12 +69,6 @@ export function TransactionListWithBalances({
   onOpenTransaction,
   onRefresh,
 }) {
-  const newTransactions = useSelector(state => state.queries.newTransactions);
-
-  const isNewTransaction = id => {
-    return newTransactions.includes(id);
-  };
-
   const selectedInst = useSelected('transactions', transactions);
 
   return (
@@ -111,7 +104,6 @@ export function TransactionListWithBalances({
         <TransactionList
           isLoading={isLoading}
           transactions={transactions}
-          isNewTransaction={isNewTransaction}
           onLoadMore={onLoadMore}
           onOpenTransaction={onOpenTransaction}
         />
